requirements.txt
uwsgi==2.0.18
mysqlclient==1.4.6
django==3.0.2
django-debug-toolbar==2.1
django-widget-tweaks==1.4.5
django-bootstrap4==1.1.1
django-model-utils==4.0.0
django-ipware==2.1.0
django-filter==2.2.0
django-active-link==0.1.6
django-slack==5.14.3
pandas
coverage
django-tempus-dominus
pre-commit
colorlog
